+++ packages/kit/src/cli.js
@@ -1,3 +1,4 @@
+import { existsSync } from 'fs';
 import sade from 'sade';
 import colors from 'kleur';
 import { load_config } from './api/load_config';
@@ -9,8 +10,14 @@ function get_config() {
 	} catch (error) {
 		let message = error.message;
 
-		if (error.code === 'ENOENT') {
-			message = 'Missing svelte.config.cjs';
+		if (error.code === 'MODULE_NOT_FOUND') {
+			if (existsSync('svelte.config.js')) {
+				// TODO this is temporary, for the benefit of early adopters
+				message =
+					'You must rename svelte.config.js to svelte.config.cjs, and snowpack.config.js to snowpack.config.cjs';
+			} else {
+				message = 'Missing svelte.config.cjs';
+			}
 		} else if (error.name === 'SyntaxError') {
 			message = 'Malformed svelte.config.cjs';
 		}